Updates1025
This update makes two small changes: (1) light adjustments to the data interface to simplify integration with the published dataset; and (2) a minor model correction: an optional activation function added to the final layer that was previously missing.

Note: the final activation is optional and does not produce significant effect on model performance. However, the numbers reported in the paper were obtained with this activation enabled.